wrote InputState impl
This will in theory be one of the main methods for querying + handling
input (the other is to listen + react directly to events).

This is implemented as a ~1kb POD data structure w/ 2 frames worth of
data on all connected input devices, and a brief event history that can
be crawled through.

So far this is just the AppClient interface; backend TBD.