Came up with a simpler and much better API built off of PODs + the KISS principle
ie. the simpler and more universal our api, the better.

The primary change here is we've abandoned the idea of having an event
queue + event dispatch using variants + event listeners (or something),
which is annoyingly complex to write and/or typesafe not typesafe.

So instead of events we'll just offer a double-buffered view of input +
window state, and detect + apply changes on eg. window. I think this is
much simpler and actually more desirable for game prototyping, so yay.

For context, this is a facade API where multithreaded modules (each
running on their own thread) are given an abstracted view of GLFW
window + event state.

Implementation TBD.